backend/requirements.txt
fastapi
uvicorn
openai
apify-client
python-dotenv
sse-starlette 